+++ readWavFuncs.h
@@ -18,7 +18,7 @@
 
 void readHeader(t_wavhdr* header, FILE* f_in);
 
-void readData(float* data_portion, int N, FILE* f_in );
+void readData(double* data_portion, int N, FILE* f_in );
 
 void calcPortions(int* portions_info, t_wavhdr* header, int N);
